+++ Realm/RLMObject_Private.h
@@ -67,6 +67,9 @@ NS_ASSUME_NONNULL_BEGIN
 // Consumes the row, so can only usefully be called once.
 @property (nonatomic, readonly) RLMObjectBase *object;
 
+// Cannot be called if `object` has already been called.
+@property (nonatomic, readonly) RLMObjectBase *objectWithoutConsumingRow;
+
 @end
 
 // Calls valueForKey: and re-raises NSUndefinedKeyExceptions