+++ RealmSwift/LinkingObjects.swift
@@ -27,9 +27,9 @@ public class LinkingObjectsBase: NSObject, NSFastEnumeration {
     internal let objectClassName: String
     internal let propertyName: String
 
-    private var cachedRLMResults: RLMResults<RLMObject>?
-    @objc private var object: RLMWeakObjectHandle?
-    @objc private var property: RLMProperty?
+    fileprivate var cachedRLMResults: RLMResults<RLMObject>?
+    @objc fileprivate var object: RLMWeakObjectHandle?
+    @objc fileprivate var property: RLMProperty?
 
     internal var rlmResults: RLMResults<RLMObject> {
         if cachedRLMResults == nil {
@@ -73,8 +73,10 @@ public class LinkingObjectsBase: NSObject, NSFastEnumeration {
  `LinkingObjects` can only be used as a property on `Object` models. Properties of this type must be declared as `let`
  and cannot be `dynamic`.
  */
-public final class LinkingObjects<T: Object>: LinkingObjectsBase {
-    /// The type of the objects represented by the linking objects.
+public final class LinkingObjects<T: Object>: LinkingObjectsBase, ThreadConfined {
+    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
+
+    /// The type of the elements contained within the collection.
     public typealias Element = T
 
     // MARK: Properties
@@ -93,6 +95,11 @@ public final class LinkingObjects<T: Object>: LinkingObjectsBase {
     public var count: Int { return Int(rlmResults.count) }
 
     // MARK: Initializers
+
+    fileprivate init(property propertyName: String) {
+        let className = (T.self as Object.Type).className()
+        super.init(fromClassName: className, property: propertyName)
+    }
 
     /**
      Creates an instance of a `LinkingObjects`. This initializer should only be called when declaring a property on a
@@ -379,7 +386,7 @@ public final class LinkingObjects<T: Object>: LinkingObjectsBase {
     }
 }
 
-extension LinkingObjects : RealmCollection {
+extension LinkingObjects: RealmCollection {
     // MARK: Sequence Support
 
     /// Returns an iterator that yields successive elements in the linking objects.
@@ -413,6 +420,39 @@ extension LinkingObjects : RealmCollection {
             return rlmResults.addNotificationBlock { _, change, error in
                 block(RealmCollectionChange.fromObjc(value: anyCollection, change: change, error: error))
             }
+    }
+}
+
+private struct LinkingObjectsHandoverMetadata {
+    var propertyName: String
+    var property: RLMProperty?
+}
+
+extension LinkingObjects: _ThreadConfined {
+    var bridgedData: RLMThreadConfined {
+        return cachedRLMResults ?? unsafeBitCast(object!.objectWithoutConsumingRow, to: Object.self).bridgedData
+    }
+
+    var bridgedMetadata: Any? {
+        return LinkingObjectsHandoverMetadata(
+            propertyName: propertyName,
+            property: property
+        )
+    }
+
+    static func bridge(data: RLMThreadConfined, metadata: Any?) -> LinkingObjects {
+        let metadata = metadata as! LinkingObjectsHandoverMetadata
+        let bridged = LinkingObjects(property: metadata.propertyName)
+        switch data {
+        case let results as RLMResults<RLMObject>:
+            bridged.cachedRLMResults = results
+        case let object as RLMObjectBase:
+            bridged.object = RLMWeakObjectHandle(object: object)
+            bridged.property = metadata.property
+        default:
+            fatalError("Unexpected handover type \(type(of: data))")
+        }
+        return bridged
     }
 }
 
@@ -504,7 +544,9 @@ public class LinkingObjectsBase: NSObject, NSFastEnumeration {
  `LinkingObjects` can only be used as a property on `Object` models. Properties of this type must
  be declared as `let` and cannot be `dynamic`.
  */
-public final class LinkingObjects<T: Object>: LinkingObjectsBase {
+public final class LinkingObjects<T: Object>: LinkingObjectsBase, ThreadConfined {
+    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
+
     /// The element type contained in this collection.
     public typealias Element = T
 
@@ -524,6 +566,11 @@ public final class LinkingObjects<T: Object>: LinkingObjectsBase {
     public var count: Int { return Int(rlmResults.count) }
 
     // MARK: Initializers
+
+    private init(property propertyName: String) {
+        let className = (T.self as Object.Type).className()
+        super.init(fromClassName: className, property: propertyName)
+    }
 
     /**
      Creates an instance of a `LinkingObjects`. This initializer should only be called when declaring a property on a
@@ -821,6 +868,39 @@ extension LinkingObjects: RealmCollectionType {
             return rlmResults.addNotificationBlock { _, change, error in
                 block(RealmCollectionChange.fromObjc(anyCollection, change: change, error: error))
             }
+    }
+}
+
+private struct LinkingObjectsHandoverMetadata {
+    var propertyName: String
+    var property: RLMProperty?
+}
+
+extension LinkingObjects: _ThreadConfined {
+    var bridgedData: RLMThreadConfined {
+        return cachedRLMResults ?? unsafeBitCast(object!.objectWithoutConsumingRow, Object.self).bridgedData
+    }
+
+    var bridgedMetadata: Any? {
+        return LinkingObjectsHandoverMetadata(
+            propertyName: propertyName,
+            property: property
+        )
+    }
+
+    static func bridge(data: RLMThreadConfined, metadata: Any?) -> LinkingObjects {
+        let metadata = metadata as! LinkingObjectsHandoverMetadata
+        let bridged = LinkingObjects(property: metadata.propertyName)
+        switch data {
+        case let results as RLMResults:
+            bridged.cachedRLMResults = results
+        case let object as RLMObjectBase:
+            bridged.object = RLMWeakObjectHandle(object: object)
+            bridged.property = metadata.property
+        default:
+            fatalError("Unexpected handover type \(data.dynamicType)")
+        }
+        return bridged
     }
 }