+++ RealmSwift/List.swift
@@ -54,7 +54,8 @@ public class ListBase: RLMListBase {
 
  Properties of `List` type defined on `Object` subclasses must be declared as `let` and cannot be `dynamic`.
  */
-public final class List<T: Object>: ListBase {
+public final class List<T: Object>: ListBase, ThreadConfined {
+    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
 
     /// The type of the elements contained within the collection.
     public typealias Element = T
@@ -469,7 +470,7 @@ public final class List<T: Object>: ListBase {
     }
 }
 
-extension List : RealmCollection, RangeReplaceableCollection {
+extension List: RealmCollection, RangeReplaceableCollection {
     // MARK: Sequence Support
 
     /// Returns a `RLMIterator` that yields successive elements in the `List`.
@@ -556,6 +557,20 @@ extension List {
     public func average<U: AddableType>(_ property: String) -> U? { fatalError() }
 }
 
+extension List: _ThreadConfined {
+    var bridgedData: RLMThreadConfined {
+        return _rlmArray
+    }
+
+    var bridgedMetadata: Any? {
+        return nil
+    }
+
+    static func bridge(data: RLMThreadConfined, metadata: Any?) -> List {
+        return List(rlmArray: data as! RLMArray)
+    }
+}
+
 #else
 
 /// :nodoc:
@@ -590,7 +605,8 @@ public class ListBase: RLMListBase {
 
  Properties of `List` type defined on `Object` subclasses must be declared as `let` and cannot be `dynamic`.
 */
-public final class List<T: Object>: ListBase {
+public final class List<T: Object>: ListBase, ThreadConfined {
+    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
 
     /// The type of the elements contained within the collection.
     public typealias Element = T
@@ -1039,6 +1055,20 @@ extension List: RealmCollectionType, RangeReplaceableCollectionType {
         return _rlmArray.addNotificationBlock { _, change, error in
             block(RealmCollectionChange.fromObjc(anyCollection, change: change, error: error))
         }
+    }
+}
+
+extension List: _ThreadConfined {
+    var bridgedData: RLMThreadConfined {
+        return _rlmArray
+    }
+
+    var bridgedMetadata: Any? {
+        return nil
+    }
+
+    static func bridge(data: RLMThreadConfined, metadata: Any?) -> List {
+        return List(rlmArray: data as! RLMArray)
     }
 }