+++ RealmSwift/Realm.swift
@@ -605,6 +605,25 @@ public final class Realm {
         try rlmRealm.writeCopy(to: fileURL, encryptionKey: encryptionKey)
     }
 
+    // MARK: Handover
+
+    /**
+     Creates a `ThreadHandover` from the passed in objects that can be safely imported on a separate thread.
+
+     Used to pass objects between threads in a thread-safe manner.
+
+     - warning: Note that the returned `ThreadHandover` must be imported at most once, and that the current version
+                of the Realm will remain pinned until this package is imported or deallocated.
+
+     - parameter objects: The objects to pass between threads.
+     - returns: A package that can be imported on a separate thread.
+
+     - seealso: `ThreadHandover`
+     */
+    public func exportThreadHandover<T: ThreadConfined>(containing objects: [T]) -> ThreadHandover<T> {
+        return ThreadHandover(realm: self, objects: objects)
+    }
+
     // MARK: Internal
 
     internal var rlmRealm: RLMRealm
@@ -654,7 +673,6 @@ extension Realm {
 
 /// The type of a block to run for notification purposes when the data in a Realm is modified.
 public typealias NotificationBlock = (_ notification: Realm.Notification, _ realm: Realm) -> Void
-
 
 // MARK: Unavailable
 
@@ -1267,6 +1285,25 @@ public final class Realm {
      */
     public func writeCopyToURL(fileURL: NSURL, encryptionKey: NSData? = nil) throws {
         try rlmRealm.writeCopyToURL(fileURL, encryptionKey: encryptionKey)
+    }
+
+    // MARK: Handover
+
+    /**
+     Creates a `ThreadHandover` from the passed in objects that can be safely imported on a separate thread.
+
+     Used to pass objects between threads in a thread-safe manner.
+     
+     - warning: Note that the returned `ThreadHandover` must be imported at most once, and that the current version
+                of the Realm will remain pinned until this package is imported or deallocated.
+
+     - parameter objects: The objects to pass between threads.
+     - returns: A package that can be imported on a separate thread.
+
+     - seealso: `ThreadHandover`
+     */
+    public func exportThreadHandover<T: ThreadConfined>(containing objects: [T]) -> ThreadHandover<T> {
+        return ThreadHandover(realm: self, objects: objects)
     }
 
     // MARK: Internal