+++ RealmSwift/RealmCollection.swift
@@ -124,7 +124,8 @@ public enum RealmCollectionChange<T> {
 /**
  A homogenous collection of `Object`s which can be retrieved, filtered, sorted, and operated upon.
 */
-public protocol RealmCollection: RandomAccessCollection, LazyCollectionProtocol, CustomStringConvertible {
+public protocol RealmCollection: RandomAccessCollection, LazyCollectionProtocol, CustomStringConvertible, ThreadConfined {
+    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
 
     /// The type of the objects contained in the collection.
     associatedtype Element: Object
@@ -367,7 +368,7 @@ public protocol RealmCollection: RandomAccessCollection, LazyCollectionProtocol,
     func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) -> NotificationToken
 }
 
-private class _AnyRealmCollectionBase<T: Object> {
+private class _AnyRealmCollectionBase<T: Object>: _ThreadConfined {
     typealias Wrapper = AnyRealmCollection<Element>
     typealias Element = T
     var realm: Realm? { fatalError() }
@@ -396,6 +397,9 @@ private class _AnyRealmCollectionBase<T: Object> {
     func setValue(_ value: Any?, forKey key: String) { fatalError() }
     func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { fatalError() }
+    var bridgedData: RLMThreadConfined { fatalError() }
+    var bridgedMetadata: Any? { fatalError() }
+    class func bridge(data: RLMThreadConfined, metadata: Any?) -> Self { fatalError() }
 }
 
 private final class _AnyRealmCollection<C: RealmCollection>: _AnyRealmCollectionBase<C.Element> {
@@ -500,6 +504,21 @@ private final class _AnyRealmCollection<C: RealmCollection>: _AnyRealmCollection
     /// :nodoc:
     override func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { return base._addNotificationBlock(block) }
+
+    // MARK: Handover
+
+    override var bridgedData: RLMThreadConfined {
+        return base.backing.bridgedData
+    }
+
+    override var bridgedMetadata: Any? {
+        return base.backing.bridgedMetadata
+    }
+
+    static override func bridge(data: RLMThreadConfined, metadata: Any?) -> _AnyRealmCollection {
+        let bridgedBase = C.backing.bridge(data: data, metadata: metadata)
+        return _AnyRealmCollection(base: bridgedBase as! C)
+    }
 }
 
 /**
@@ -514,7 +533,13 @@ public final class AnyRealmCollection<T: Object>: RealmCollection {
 
     /// The type of the objects contained in the collection.
     public typealias Element = T
-    private let base: _AnyRealmCollectionBase<T>
+    fileprivate let base: _AnyRealmCollectionBase<T>
+
+    // MARK: Initializers
+
+    fileprivate init(base: _AnyRealmCollectionBase<T>) {
+        self.base = base
+    }
 
     /// Creates an `AnyRealmCollection` wrapping `base`.
     public init<C: RealmCollection>(_ base: C) where C.Element == T {
@@ -803,6 +828,31 @@ public final class AnyRealmCollection<T: Object>: RealmCollection {
         -> NotificationToken { return base._addNotificationBlock(block) }
 }
 
+// MARK: ThreadConfined
+
+private struct AnyRealmCollectionThreadConfinedMetadata<T: Object> {
+    var baseMetadata: Any?
+    var baseType: _AnyRealmCollectionBase<T>.Type
+}
+
+extension AnyRealmCollection: _ThreadConfined {
+    var bridgedData: RLMThreadConfined {
+        return base.bridgedData
+    }
+
+    var bridgedMetadata: Any? {
+        return AnyRealmCollectionThreadConfinedMetadata(
+            baseMetadata: base.bridgedMetadata,
+            baseType: type(of: base)
+        )
+    }
+
+    static func bridge(data: RLMThreadConfined, metadata: Any?) -> AnyRealmCollection {
+        let metadata = metadata as! AnyRealmCollectionThreadConfinedMetadata<T>
+        let bridgedBase = metadata.baseType.bridge(data: data, metadata: metadata.baseMetadata)
+        return AnyRealmCollection(base: bridgedBase)
+    }
+}
 
 // MARK: Unavailable
 
@@ -939,7 +989,8 @@ public enum RealmCollectionChange<T> {
  A homogenous collection of `Object`s which can be retrieved, filtered, sorted,
  and operated upon.
 */
-public protocol RealmCollectionType: CollectionType, CustomStringConvertible {
+public protocol RealmCollectionType: CollectionType, CustomStringConvertible, ThreadConfined {
+    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
 
     /// The type of the objects contained in the collection.
     associatedtype Element: Object
@@ -1171,7 +1222,7 @@ public protocol RealmCollectionType: CollectionType, CustomStringConvertible {
     func _addNotificationBlock(block: (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) -> NotificationToken
 }
 
-private class _AnyRealmCollectionBase<T: Object> {
+private class _AnyRealmCollectionBase<T: Object>: _ThreadConfined {
     typealias Wrapper = AnyRealmCollection<Element>
     typealias Element = T
     var realm: Realm? { fatalError() }
@@ -1200,6 +1251,9 @@ private class _AnyRealmCollectionBase<T: Object> {
     func setValue(value: AnyObject?, forKey key: String) { fatalError() }
     func _addNotificationBlock(block: (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { fatalError() }
+    var bridgedData: RLMThreadConfined { fatalError() }
+    var bridgedMetadata: Any? { fatalError() }
+    class func bridge(data: RLMThreadConfined, metadata: Any?) -> Self { fatalError() }
 }
 
 private final class _AnyRealmCollection<C: RealmCollectionType>: _AnyRealmCollectionBase<C.Element> {
@@ -1294,6 +1348,21 @@ private final class _AnyRealmCollection<C: RealmCollectionType>: _AnyRealmCollec
     /// :nodoc:
     override func _addNotificationBlock(block: (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { return base._addNotificationBlock(block) }
+
+    // MAKR: Handover
+
+    override var bridgedData: RLMThreadConfined {
+        return base.backing.bridgedData
+    }
+
+    override var bridgedMetadata: Any? {
+        return base.backing.bridgedMetadata
+    }
+
+    static override func bridge(data: RLMThreadConfined, metadata: Any?) -> _AnyRealmCollection {
+        let bridgedBase = C.backing.bridge(data, metadata: metadata)
+        return _AnyRealmCollection(base: bridgedBase as! C)
+    }
 }
 
 /**
@@ -1302,11 +1371,18 @@ private final class _AnyRealmCollection<C: RealmCollectionType>: _AnyRealmCollec
  Instances of `RealmCollectionType` forward operations to an opaque underlying collection having the same `Element`
  type.
  */
-public final class AnyRealmCollection<T: Object>: RealmCollectionType {
+public final class AnyRealmCollection<T: Object>: RealmCollectionType, ThreadConfined {
+    // FIXME: Remove redundant conformance to `ThreadConfined` once bug SR-2146 is fixed.
 
     /// The type of the objects contained in the collection.
     public typealias Element = T
     private let base: _AnyRealmCollectionBase<T>
+
+    // MARK: Initializers
+
+    private init(base: _AnyRealmCollectionBase<T>) {
+        self.base = base
+    }
 
     /// Creates an `AnyRealmCollection` wrapping `base`.
     public init<C: RealmCollectionType where C.Element == T>(_ base: C) {
@@ -1572,6 +1648,32 @@ public final class AnyRealmCollection<T: Object>: RealmCollectionType {
     /// :nodoc:
     public func _addNotificationBlock(block: (RealmCollectionChange<AnyRealmCollection>) -> Void)
         -> NotificationToken { return base._addNotificationBlock(block) }
+}
+
+// MARK: ThreadConfined
+
+private struct AnyRealmCollectionThreadConfinedMetadata<T: Object> {
+    var baseMetadata: Any?
+    var baseType: _AnyRealmCollectionBase<T>.Type
+}
+
+extension AnyRealmCollection: _ThreadConfined {
+    var bridgedData: RLMThreadConfined {
+        return base.bridgedData
+    }
+
+    var bridgedMetadata: Any? {
+        return AnyRealmCollectionThreadConfinedMetadata(
+            baseMetadata: base.bridgedMetadata,
+            baseType: base.dynamicType
+        )
+    }
+
+    static func bridge(data: RLMThreadConfined, metadata: Any?) -> AnyRealmCollection {
+        let metadata = metadata as! AnyRealmCollectionThreadConfinedMetadata<T>
+        let bridgedBase = metadata.baseType.bridge(data, metadata: metadata.baseMetadata)
+        return AnyRealmCollection(base: bridgedBase)
+    }
 }
 
 #endif