+++ RealmSwift/Tests/PerformanceTests.swift
@@ -340,7 +340,7 @@ class SwiftPerformanceTests: TestCase {
 
     func testRealmCreationCached() {
         var realm: Realm!
-        dispatchSyncNewThread {
+        dispatchAsyncAndWait {
             realm = try! Realm()
         }
 
@@ -806,7 +806,7 @@ class SwiftPerformanceTests: TestCase {
 
     func testRealmCreationCached() {
         var realm: Realm!
-        dispatchSyncNewThread {
+        dispatchAsyncAndWait {
             realm = try! Realm()
         }