+++ RealmSwift/Tests/TestCase.swift
@@ -106,7 +106,7 @@ class TestCase: XCTestCase {
         RLMRealm.resetRealmState()
     }
 
-    func dispatchSyncNewThread(block: @escaping () -> Void) {
+    func dispatchAsyncAndWait(block: @escaping () -> Void) {
         queue.async {
             autoreleasepool {
                 block()
@@ -200,6 +200,8 @@ class TestCase: XCTestCase {
     var exceptionThrown = false
     var testDir: String! = nil
 
+    let queue = dispatch_queue_create("background", nil)
+
     func realmWithTestPath(configuration: Realm.Configuration = Realm.Configuration()) -> Realm {
         var configuration = configuration
         configuration.fileURL = testRealmURL()
@@ -269,8 +271,7 @@ class TestCase: XCTestCase {
         RLMRealm.resetRealmState()
     }
 
-    func dispatchSyncNewThread(block: dispatch_block_t) {
-        let queue = dispatch_queue_create("background", nil)
+    func dispatchAsyncAndWait(block: dispatch_block_t) {
         dispatch_async(queue) {
             autoreleasepool {
                 block()